pyproject.toml
[project]
name = "aiiac"
version = "0.1.0"
description = "AI Infrastructure as Code Generator"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "click>=8.0.0",
    "rich>=10.0.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.10",
    "pydantic>=2.0.0",
    "requests>=2.0.0",
    "PyYAML>=6.0.0",
]
requires-python = ">=3.9"

[project.scripts]
aiiac = "aiiac.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"